requirements.txt
ynab>=1.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
openpyxl>=3.0.7
reportlab>=3.6.1
schedule>=1.1.0
requests>=2.26.0
rich>=10.0.0
tabulate>=0.8.9
flask>=2.0.0
plotly>=5.3.0
flask-cors>=3.0.0 